refactor(RefreshableViewModel): NetworkViewModel에서 RefreshableViewModel 분리
