+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/base/NetworkViewModel.kt
@@ -42,8 +42,6 @@ abstract class NetworkViewModel : ViewModel() {
         changeToLoadingState()
     }
 
-    abstract fun refresh(): Job
-
     protected fun <T : Any> requestNetwork(
         request: suspend () -> ApiResponse<T>,
         onSuccess: suspend (T) -> Unit,
@@ -74,42 +72,6 @@ abstract class NetworkViewModel : ViewModel() {
         onFinish()
     }
 
-    protected fun <T : Any> fetchData(
-        fetchData: suspend () -> ApiResponse<T>,
-        onSuccess: (T) -> Unit = {},
-        onFailure: (code: Int, message: String?) -> Unit = { _, _ -> },
-        onLoading: suspend () -> Unit = { changeToLoadingState() },
-        onNetworkError: () -> Unit = ::changeToNetworkErrorState,
-        onStart: () -> Unit = {},
-        onFinish: () -> Unit = {},
-    ): Job = requestNetwork(
-        request = { fetchData() },
-        onSuccess = { onSuccess(it) },
-        onFailure = { code, message -> onFailure(code, message) },
-        onLoading = { onLoading() },
-        onNetworkError = { onNetworkError() },
-        onStart = { onStart() },
-        onFinish = { onFinish() },
-    )
-
-    protected fun <T : Any> refreshData(
-        refresh: suspend () -> ApiResponse<T>,
-        onSuccess: (T) -> Unit = {},
-        onFailure: (code: Int, message: String?) -> Unit = { _, _ -> },
-        onLoading: suspend () -> Unit = {},
-        onNetworkError: () -> Unit = {},
-        onStart: () -> Unit = {},
-        onFinish: () -> Unit = {},
-    ): Job = requestNetwork(
-        request = { refresh() },
-        onSuccess = { onSuccess(it) },
-        onFailure = { code, message -> onFailure(code, message) },
-        onLoading = { onLoading() },
-        onNetworkError = { onNetworkError() },
-        onStart = { onStart() },
-        onFinish = { onFinish() },
-    )
-
     protected fun <T : Any> command(
         command: suspend () -> ApiResponse<T>,
         onSuccess: (T) -> Unit = {},
@@ -121,28 +83,6 @@ abstract class NetworkViewModel : ViewModel() {
     ): Job = requestNetwork(
         request = { command() },
         onSuccess = { onSuccess(it) },
-        onFailure = { code, message -> onFailure(code, message) },
-        onLoading = { onLoading() },
-        onNetworkError = { onNetworkError() },
-        onStart = { onStart() },
-        onFinish = { onFinish() },
-    )
-
-    protected fun <T : Any> commandAndRefresh(
-        command: suspend () -> ApiResponse<T>,
-        onSuccess: suspend (T) -> Unit = {},
-        onFailure: (code: Int, message: String?) -> Unit = { _, _ -> },
-        onLoading: suspend () -> Unit = { delayLoading() },
-        onNetworkError: () -> Unit = ::dispatchNetworkErrorEvent,
-        onStart: () -> Unit = {},
-        onFinish: () -> Unit = {},
-        refresh: () -> Job = { this@NetworkViewModel.refresh() },
-    ): Job = requestNetwork(
-        request = { command() },
-        onSuccess = {
-            refresh().join()
-            onSuccess(it)
-        },
         onFailure = { code, message -> onFailure(code, message) },
         onLoading = { onLoading() },
         onNetworkError = { onNetworkError() },
