+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/recruitmentList/EventRecruitmentViewModel.kt
@@ -3,7 +3,7 @@ package com.emmsale.presentation.ui.recruitmentList
 import androidx.lifecycle.SavedStateHandle
 import com.emmsale.data.model.Recruitment
 import com.emmsale.data.repository.interfaces.RecruitmentRepository
-import com.emmsale.presentation.base.NetworkViewModel
+import com.emmsale.presentation.base.RefreshableViewModel
 import com.emmsale.presentation.common.livedata.NotNullLiveData
 import com.emmsale.presentation.common.livedata.NotNullMutableLiveData
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -14,7 +14,7 @@ import javax.inject.Inject
 class EventRecruitmentViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
     private val recruitmentRepository: RecruitmentRepository,
-) : NetworkViewModel() {
+) : RefreshableViewModel() {
     val eventId: Long = requireNotNull(savedStateHandle[EVENT_ID_KEY]) {
         "[ERROR] 행사 아이디를 가져오지 못했어요"
     }
