+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/scrappedEventList/ScrappedEventViewModel.kt
@@ -2,7 +2,7 @@ package com.emmsale.presentation.ui.scrappedEventList
 
 import com.emmsale.data.model.Event
 import com.emmsale.data.repository.interfaces.EventRepository
-import com.emmsale.presentation.base.NetworkViewModel
+import com.emmsale.presentation.base.RefreshableViewModel
 import com.emmsale.presentation.common.livedata.NotNullLiveData
 import com.emmsale.presentation.common.livedata.NotNullMutableLiveData
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -12,7 +12,7 @@ import javax.inject.Inject
 @HiltViewModel
 class ScrappedEventViewModel @Inject constructor(
     private val eventRepository: EventRepository,
-) : NetworkViewModel() {
+) : RefreshableViewModel() {
     private val _scrappedEvents = NotNullMutableLiveData(listOf<Event>())
     val scrappedEvents: NotNullLiveData<List<Event>> = _scrappedEvents
